_tests/browser_testing/requirements.txt
selenium==2.44.0
pyhamcrest==1.8.1
behave==1.2.4